refactor: derive client-backed cli routing
