+++ src/cli/commands/__tests__/generic.test.ts
@@ -3,7 +3,7 @@ import assert from 'node:assert/strict';
 import { createAgentDeviceClient } from '../../../client/client.ts';
 import type { DaemonResponse } from '../../../kernel/contracts.ts';
 import type { CliFlags } from '../../parser/cli-flags.ts';
-import type { ClientBackedCliCommandName } from '../../../command-catalog.ts';
+import type { ClientBackedCliCommandName } from '../client-backed.ts';
 import { runGenericClientBackedCommand } from '../generic.ts';
 
 async function captureStdout(fn: () => Promise<unknown>): Promise<string> {
